requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.3.1
six==1.17.0
soupsieve==2.7
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
